requirements.txt
grpcio==1.44.0
grpcio-tools==1.44.0
awscli==1.22.49
s3transfer==0.5.2
boto3~=1.19.6
pycurl~=7.44.1
requests~=2.26.0
botocore~=1.23.49
jmespath~=0.10.0
google~=3.0.0
cryptography~=35.0.0
six~=1.16.0
python-dateutil~=2.8.2
urllib3~=1.26.7
certifi~=2021.10.8
aiocurl~=0.0.3.post1
tqdm~=4.62.3
idna~=3.3
beautifulsoup4~=4.11.1
lxml~=4.6.3
soupsieve~=2.3.2.post1
numpy~=1.21.3
setuptools~=57.0.0
docutils~=0.15.2